+++ application/views/pages/admin/navigation.php
@@ -6,7 +6,7 @@
 		<?php
 		if (isset($_GET['id']))
 			{
-			$creative_id = $_GET['id'];
+		$creative_id = $_GET['id'];
 		echo '<li><div>';
 		echo	'<div class="">' ; 
 		echo	"<a href='creative_edit?id=$creative_id'>Edit</a> |"; 
@@ -14,7 +14,8 @@
 		echo	"<a href='creative_markets?id=$creative_id'>Markets</a> |"; 
 		echo	"<a href='creative_categories?id=$creative_id'>Categories</a> | ";
 		echo	"<a href='creative_tags?id=$creative_id'>Tags</a> | ";
-		echo	"<a href='creative_delete?id=$creative_id'>Delete</a> ";
+		/*echo	"<a href='creative_delete?id=$creative_id'>Delete</a> ";*/
+		echo	"<a href='creative_delete?id=$creative_id'onclick= \"return confirm('Do you really want to delete?');\" >Delete</a> ";
 		echo	'</div>';
 		echo '</div></li>';
